A big revamp of keymap parsing.
This introduces the ability to reference fns from other modes. For example:
meta:find-file references the find-file fn defined in meta-mode.

It also revamps keymap parsing to use Either (and to slightly improve error
messages). It's still a pig's breakfast in there, but I can't seem to find a
way to express more clearly or elegantly a computation that can fail six ways
to Sunday. I tried.