Revamped Indenter system.
Instead of relying on a bunch of ad-hoc regexp matching and searching around
the buffer, we do a simple form of parsing and assign indentation states to
each line of the buffer. The combination of the current indentation state and
the contents of a line should be enough to compute a line's indentation.

If information is needed from elsewhere in the buffer to determine the correct
indentation for a given line, it needs to be encoded into the indentation state
and calculated along with the rest.

This makes some things a little fiddlier, but massively simplifies a bunch of
other stuff.