Omit Config.Defs.global; added some Java helpers.
Scala helpfully erases generics to Object when instantiated as primitives,
instead of to their corresponding box type. This makes them effectively
unusable for Java. Yay. So we introduce bindings directly to the Java box
types, which can be used from Java without major pain. And naturally in this
case by Java, I mean any JVM language that is not Scala.