Revamped SubProcess to be event based.
This also includes events to indicate that stdout and stderr have "completed"
(we detected EOF on them).

Lacking those made it difficult to know when your subprocess was done
generating output. Using waitFor was insufficient because that could return
before the final output was read from the process (by a background thread).