More visit-y goodness.
VisitList -> Visit.List, VisitStack -> Visit.Stack. Made the Visit.Stack into a
standard thing as well and integrated it with the visit list. So when you start
wandering off through the visit list, we push your old location onto the stack
so that you can easily pop back to it.