Tweak how auto-quote/bracket pairing is done.
We now only "overwrite" a close quote/bracket if it's tagged as being
auto-inserted.

I think I can smarten up the check for whether we should auto-insert at all
because right now it's kind of conservative, but this should at least put an
end to all of the annoying unintended close overwrites.