Allow minor modes to be triggered by state types.
For example: if we stick a Project instance into buffer state, trigger
project-mode. Many minor modes require a piece of state in the buffer to tell
them what they're doing, might as well trigger the mode directly based on that
state instead of requiring that one jam a tag in there too. Simplify!