Further investigation of slowness on big files.
It looks like JavaFX just shits itself if we add a lot of (even simple)
children to a Region. Even if those children are invisible. So we're going to
have to revamp BufferArea to remove line views that are not visible and add
those that are. This is annoying because we have to cope with line views being
inserted into the visible area, but we'll work it out in the name of not being
fucking slow when you open a 30k line file.