Add BlockS.makeEOL.
This is useful in cases where a block char *thinks* it's not at the end of the
line so it shouldn't cause an indent, but in reality it is the last important
thing on the line, so it should.

The mode in question can identify said circumstance and convert that BlockS
into one that does trigger indent.