Only ever add visible lines to scene graph.
We used to add all lines to scene graph and toggle their visibility. But JavaFX
isn't very good at handling hundreds of thousands of nodes, so this failed
miserably on large text files.

Now we only ever have nodes in the scene graph for the currently visible lines,
which never numbers more than a few hundred even on the most insanely large
monitors.

We could go even further and not even *have* nodes for non-visible lines, which
would probably be just fine and save a bunch of memory, but I don't want to go
there today because this was a short diversion, and I'm not 100% sure I want to
go there at all, because that would impact potential future support for
"annotating" the scene graph for lines.

Maybe that's fine and we'll always structure that such that annotations are
generated on demand when a line becomes visible, but maybe that's a PITA and we
should not prematurely optimize for memory savings on massive files. Anyway...