Move ` and ^ to punctuation class.
They're classified by Unicode as modifier symbols like the other
combining accents, which makes them words. But they're in ASCII and used
by programs as punctuation, so we're going to err on the side of making
that DWIM, at the potential expense of languages where these might
somehow appear in words... Sorry world.