A wee collections rethink.
Iterable gains withFilter and flatMap. Option is no longer implicitly an
Iterable but can be viewed as one (implicitly if you're programmin' in Scala).

I think I was using the old approach because I wanted to be able to specialize
the map/flatMap/etc. methods on Option but still allow it to be Unordered and
not have crazy override conflicts.

This is simpler and doesn't lobotomize the Iterable API to smarten up the
Option API. It does mean we can't rely on eta expansion when using method that
returns option in a map/flatMap chain on iterables, but I'm going to call that
an acceptable sacrifice. I had to fix exactly one such usage in the entire
Scaled code base.

This also makes all (Iterable-descended) flatMaps take A => JIterable[A], which
only some of them did before (god knows what the type checker thought of it
all). This should ostensibly improve Java interop.