Revamp code completion UX.
We now work more like VS Code and other IDEs: when you type a trigger
character, we automatically compute and show a completion and as you type
additional characters the completion is refined. You can ignore it, or press
tab or enter to accept it.

I still need to add support for showing completions when the trigger character
was not pressed, but that should follow fairly readily.

I like manually initiating a completion with tab, but I have to admit that the
automatic completions with fuzzy refinement feel really good. I'm going to go
with the herd on this one.