Tweaking completion UX.
Pressing TAB will (once again) request a completion. If we have an active
completion, it will extend the completion to the longest shared prefix. If
we're already at that longest prefix, it will advance the cursor to the next
completion choice (shift-TAB moves cursor backwards).

You can press ENTER at any point to accept the highlighted completion, or keep
extending and refining your way to a single completion at which point it
considers itself done. Best of both worlds, IMO.