Give popups a way to add styles to their buffer.
In theory they could style the lines before creating the popup, but all of the
syntax highlighting infrastructure is designed to work directly on buffers, not
during the line construction process (because it happens on code as it is
edited). So it's way easier to just run that stuff on the buffer we end up
creating for the popup than to factor it out so that it can work both at line
construction time and on an already constructed buffer.