Completion improvements.
A basic completion Choice now just has the insert text. One can override Choice
to provide a specialized label to show and to provide a buffer for the details
popup which is told the width of the view so it can wrap sensibly.

More work forthcoming once I cajole the Eclipse LSP into giving me
snippet-style completions.