Consolidated and revamped block indenter.
ByBlock becomes BlockIndenter which combines state with "block"-ness, and also
breaks out the hooks into composable Rule objects. This allows us to create a
bunch of reusable rules which a given language can mix and match from based on
its special snowflake needs.

It makes it harder to handle interactions between rules, but most of the rules
are fairly orthogonal, and worst case a language can eschew one or more
standard rules and mash them together in a custom rule that handles its special
complexity.