Put the padding back.
I think I disabled this because I didn't like that after typing enter, having
the '*' of a C block comment inserted and then a space, I had to backspace to
turn '* ' into '* /'. But I think it's more annoying to have to press space
after I press enter to start typing an @command or just a manually broken line.

If I want more magic, I'll just have something that turns "* /" into "*/" if I
type / as the first thing after newline in a block or doc comment.