Don't overwrite minibuffer contents.
In general we don't want to do that except in very specific circumstances, like
extending completions with a shared prefix or twiddling case to match the
completion. We definitely don't want to do it when we hear back from an async
method because the user may have moved on and typed new things and we then wipe
those out with some old stale completion request.